refactor: separate task and timesheet forms from tables
