+++ components/table/TimesheetTable.tsx
@@ -1,49 +1,19 @@
 import { Button } from '@chakra-ui/button';
 import { Box, Flex, Heading } from '@chakra-ui/layout';
-import {
-    FormControl,
-    FormErrorMessage,
-    FormLabel,
-    Input,
-    Modal,
-    ModalCloseButton,
-    ModalContent,
-    ModalFooter,
-    ModalHeader,
-    ModalOverlay,
-    Select,
-    useDisclosure,
-} from '@chakra-ui/react';
+import { Modal, ModalCloseButton, ModalContent, ModalHeader, ModalOverlay } from '@chakra-ui/react';
 import { Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/table';
 import React, { useState } from 'react';
 import ErrorAlert from '../common/ErrorAlert';
-import { postTimesheet, putTimesheet } from '@/lib/utils/apiRequests';
+import { putTimesheet } from '@/lib/utils/apiRequests';
 import { EmployeeDTO, ProjectDTO, TimesheetDTO } from '@/lib/types/dto';
+import { TimesheetForm } from '../form/TimesheetForm';
 
-type TimesheetFields = Partial<TimesheetDTO> & {
-    project: ProjectDTO;
-};
-
-const validateTimesheetFields = (fields: TimesheetFields): TimesheetDTO => {
-    const { name, project, employee } = fields;
-    if (name && project && employee) {
-        return {
-            ...fields,
-            name,
-            project,
-            employee,
-        };
-    } else {
-        throw 'Invalid timesheet form: Missing required fields';
-    }
-};
-
-type TimesheetTableProps = {
+interface TimesheetTableProps {
     project: ProjectDTO;
     timesheets: TimesheetDTO[];
     refreshTimesheets: () => void;
     employees: EmployeeDTO[];
-};
+}
 
 function TimesheetTable({
     project,
@@ -51,20 +21,8 @@ function TimesheetTable({
     timesheets: previousTimesheets,
     employees,
 }: TimesheetTableProps): JSX.Element {
-    const [timesheetFields, setTimesheetFields] = useState<TimesheetFields>({ project });
-    const { isOpen, onOpen, onClose } = useDisclosure();
+    const [displayNewTimesheetForm, setDisplayNewTimesheetForm] = useState(false);
     const [errorMessage, setErrorMessage] = useState<string>('');
-
-    const submitTimesheet = async (e: React.MouseEvent) => {
-        e.preventDefault();
-        try {
-            await postTimesheet(validateTimesheetFields(timesheetFields));
-            await refreshTimesheets();
-            onClose();
-        } catch (error) {
-            setErrorMessage(`${error}`);
-        }
-    };
 
     const archiveTimesheet = async (timesheet: TimesheetDTO, e: React.MouseEvent) => {
         e.preventDefault();
@@ -75,22 +33,6 @@ function TimesheetTable({
             setErrorMessage(`${error}`);
         }
     };
-
-    const setEmployee = (e: React.FormEvent<HTMLSelectElement>) => {
-        e.preventDefault();
-        const id = parseInt(e.currentTarget.value);
-        if (id && employees) {
-            const employee = employees.find((employee) => employee.id === id);
-            if (employee) {
-                setTimesheetFields({ ...timesheetFields, employee: { ...employee }, project: { ...project } });
-            } else {
-                throw `Error timesheet form could not find employee with id ${id}.`;
-            }
-        }
-    };
-
-    const invalidAllocation =
-        (timesheetFields.allocation && (timesheetFields.allocation < 0 || timesheetFields.allocation > 100)) || false;
 
     return (
         <Flex
@@ -149,78 +91,21 @@ function TimesheetTable({
                 </Box>
             )}
             <Flex flexDirection="row-reverse">
-                <Button colorScheme="blue" onClick={onOpen}>
+                <Button colorScheme="blue" onClick={() => setDisplayNewTimesheetForm(true)}>
                     Add User
                 </Button>
             </Flex>
-            <Modal isOpen={isOpen} onClose={onClose}>
+            <Modal isOpen={displayNewTimesheetForm} onClose={() => setDisplayNewTimesheetForm(false)}>
                 <ModalOverlay />
                 <ModalContent px="0.5rem">
                     <ModalHeader>Add user to project</ModalHeader>
+                    <TimesheetForm
+                        project={project}
+                        employees={employees}
+                        refreshTimesheets={refreshTimesheets}
+                        onClose={() => setDisplayNewTimesheetForm(false)}
+                    />
                     <ModalCloseButton />
-                    <FormControl>
-                        <FormLabel>User</FormLabel>
-                        <Select onChange={setEmployee} placeholder="Select employee">
-                            {employees.map((employee, idx) => {
-                                return (
-                                    <option key={idx} value={employee.id}>
-                                        {`${employee.first_name} ${employee.last_name}`}
-                                    </option>
-                                );
-                            })}
-                        </Select>
-                    </FormControl>
-                    <FormControl>
-                        <FormLabel>Timesheet Name</FormLabel>
-                        <Input
-                            placeholder="Timesheet Name"
-                            onChange={(e) =>
-                                setTimesheetFields({
-                                    ...timesheetFields,
-                                    name: e.target.value,
-                                })
-                            }
-                        />
-                    </FormControl>
-                    <FormControl>
-                        <FormLabel>Description</FormLabel>
-                        <Input
-                            placeholder="Description"
-                            onChange={(e) =>
-                                setTimesheetFields({
-                                    ...timesheetFields,
-                                    description: e.target.value,
-                                })
-                            }
-                        />
-                    </FormControl>
-                    <FormControl isInvalid={invalidAllocation}>
-                        <FormLabel>Allocation</FormLabel>
-                        <Input
-                            placeholder="0"
-                            onChange={(e) =>
-                                setTimesheetFields({
-                                    ...timesheetFields,
-                                    allocation: parseInt(e.target.value),
-                                })
-                            }
-                        />
-                        <FormErrorMessage>Allocation needs to be between 1 and 100 %.</FormErrorMessage>
-                    </FormControl>
-                    <ModalFooter>
-                        <Button colorScheme="blue" mr={3} onClick={submitTimesheet}>
-                            Submit
-                        </Button>
-                        <Button colorScheme="gray" onClick={onClose}>
-                            Cancel
-                        </Button>
-                    </ModalFooter>
-                    {errorMessage ? (
-                        <>
-                            <ErrorAlert />
-                            <Box>{errorMessage}</Box>
-                        </>
-                    ) : null}
                 </ModalContent>
             </Modal>
         </Flex>
